abis: Give termios.h the abi-bits treatment. options/posix: Give termios.h the abi-bits treatment.
