+++ options/posix/include/termios.h
@@ -3,6 +3,7 @@
 #define _TERMIOS_H
 
 #include <bits/posix/pid_t.h>
+#include <abi-bits/termios.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -11,22 +12,6 @@ extern "C" {
 typedef unsigned int cc_t;
 typedef unsigned int speed_t;
 typedef unsigned int tcflag_t;
-
-// indices for the c_cc array in struct termios
-#define NCCS    11
-#define VEOF     0
-#define VEOL     1
-#define VERASE   2
-#define VINTR    3
-#define VKILL    4
-#define VMIN     5
-#define VQUIT    6
-#define VSTART   7
-#define VSTOP    8
-#define VSUSP    9
-#define VTIME   10
-#define VLNEXT  15
-#define VEOL2   16
 
 // bitwise flags for c_iflag in struct termios
 #define BRKINT  0x0001
@@ -137,16 +122,6 @@ typedef unsigned int tcflag_t;
 #define TCION 2
 #define TCOOFF 3
 #define TCOON 4
-
-struct termios {
-	tcflag_t c_iflag;
-	tcflag_t c_oflag;
-	tcflag_t c_cflag;
-	tcflag_t c_lflag;
-	cc_t c_cc[NCCS];
-	speed_t ibaud;
-	speed_t obaud;
-};
 
 speed_t cfgetispeed(const struct termios *);
 speed_t cfgetospeed(const struct termios *);
