+++ sysdeps/managarm/include/abi-bits/termios.h
@@ -1,0 +1,1 @@
+../../../../abis/linux/x86_64/termios.h
